% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.family.R
\name{graph.family}
\alias{graph.family}
\title{Visualization of PCA by taxonomic family}
\usage{
graph.family(data, family.name)
}
\arguments{
\item{data}{dataframe including numerical data as well as information on taxonomic family and genus}

\item{family.name}{family name as it appears in the dataframe}
}
\description{
This function graphs a PCA on a user-defined taxonomic family
}
\examples{
graph.family(data = d, family.name = "Cercopithecidae")
}
\keyword{PCA}
\keyword{graph}
\keyword{taxonomic}
