% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.superfamily.R
\name{graph.superfamily}
\alias{graph.superfamily}
\title{Visualization of PCA by taxonomic superfamily}
\usage{
graph.superfamily(data, superfamily.name)
}
\arguments{
\item{data}{dataframe including numerical data as well as information on taxonomic superfaimly and family}

\item{superfamily.name}{superfamily name as it appears in the dataframe}
}
\description{
This function graphs a PCA on a user-defined taxonomic superfamily
}
\examples{
graph.superfamily(data = d, superfamily.name = "Ceboidea")
}
\keyword{PCA}
\keyword{graph}
\keyword{taxonomic}
