% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonpca-package.R
\docType{package}
\name{taxonpca-package}
\alias{taxonpca}
\alias{taxonpca-package}
\title{taxonpca: Taxonomic Principle Component Analysis}
\description{
This package allows easy analysis of biological data taxonomically using PCA.
}
\author{
\strong{Maintainer}: Lydia Myers \email{lydia.myers@utexas.edu}

Authors:
\itemize{
  \item Gabrielle Bueno
}

}
\keyword{internal}
